fix pnp e2e test

+++ digitaltwin_client/tests/dt_interface_client_ut/dt_interface_client_ut.c
@@ -450,7 +450,7 @@ static void test_Impl_DT_PropertyUpdate(const DIGITALTWIN_CLIENT_PROPERTY_UPDATE
         if (!(dtTestPropertyUpdateTest == DT_TEST_PROPERTY_UPDATE_TEST_REPORTED_ONLY_EXPECTED) && (expectedDesiredPropertyData == NULL))
         {
             ASSERT_ARE_EQUAL(size_t, strlen(expectedDesiredPropertyData), dtClientPropertyUpdate->propertyDesiredLen,
-                "Expected desired property length <%lu> does not match actual <%lu>", strlen(expectedDesiredPropertyData), dtClientPropertyUpdate->propertyDesiredLen);
+                "Expected desired property length <%lu> does not match actual <%lu>", (unsigned long)(strlen(expectedDesiredPropertyData)), (unsigned long)(dtClientPropertyUpdate->propertyDesiredLen));
             ASSERT_ARE_EQUAL(int, 0, strncmp(expectedDesiredPropertyData, (char*)dtClientPropertyUpdate->propertyDesired, dtClientPropertyUpdate->propertyDesiredLen),
                 "Expected desired property data <%s> does not match actual <%.*s>", expectedPropertyName, (int)dtClientPropertyUpdate->propertyDesiredLen, (char*)dtClientPropertyUpdate->propertyDesired);
         }
@@ -465,7 +465,7 @@ static void test_Impl_DT_PropertyUpdate(const DIGITALTWIN_CLIENT_PROPERTY_UPDATE
             const char* expectedReportedPropertyData = dtTestExpectedPropertyStatus.expectedReportedData[i];
             ASSERT_IS_NOT_NULL(dtClientPropertyUpdate->propertyReported, "Reported property is NULL");
             ASSERT_ARE_EQUAL(int, strlen(expectedReportedPropertyData), dtClientPropertyUpdate->propertyReportedLen,
-                             "Expected propertyReportedLen <%lu> does not match actual <%lu>", strlen(expectedReportedPropertyData), dtClientPropertyUpdate->propertyReportedLen);
+                             "Expected propertyReportedLen <%lu> does not match actual <%lu>", (unsigned long)(strlen(expectedReportedPropertyData)), (unsigned long)(dtClientPropertyUpdate->propertyReportedLen));
             ASSERT_ARE_EQUAL(int, 0, strcmp(expectedReportedPropertyData, (char*)dtClientPropertyUpdate->propertyReported),
                              "Expected propertyReported <%s> does not match expected <%.*s>", expectedReportedPropertyData, (int)dtClientPropertyUpdate->propertyReportedLen, dtClientPropertyUpdate->propertyReported);
         }
